// packages/apps-wake/src/utils/graphql/storefront.graphql.gen.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Any: { input: any; output: any };
  CEP: { input: any; output: any };
  CountryCode: { input: any; output: any };
  DateTime: { input: any; output: any };
  Decimal: { input: any; output: any };
  EmailAddress: { input: any; output: any };
  Long: { input: any; output: any };
  Upload: { input: any; output: any };
  Uuid: { input: any; output: any };
};
// ...
export type AttributeInput = {
  /** The attribute Ids to match. */
  id?: InputMaybe<Array<Scalars["Long"]["input"]>>;
  /** The attribute name to match. */
  name?: InputMaybe<Array<InputMaybe<Scalars["String"]["input"]>>>;
  /** The attribute type to match. */
  type?: InputMaybe<Array<InputMaybe<Scalars["String"]["input"]>>>;
  /** The attribute value to match */
  value?: InputMaybe<Array<InputMaybe<Scalars["String"]["input"]>>>;
};
// ...
export type PricesInput = {
  /** The product discount must be greater than or equal to. */
  discount_gte?: InputMaybe<Scalars["Decimal"]["input"]>;
  /** The product discount must be lesser than or equal to. */
  discount_lte?: InputMaybe<Scalars["Decimal"]["input"]>;
  /** Return only products where the listed price is more than the price. */
  discounted?: InputMaybe<Scalars["Boolean"]["input"]>;
  /** The product price must be greater than or equal to. */
  price_gte?: InputMaybe<Scalars["Decimal"]["input"]>;
  /** The product price must be lesser than or equal to. */
  price_lte?: InputMaybe<Scalars["Decimal"]["input"]>;
};
// ...
export type ProductExplicitFiltersInput = {
  /** The set of attributes do filter. */
  attributes?: InputMaybe<AttributeInput>;
  /** Choose if you want to retrieve only the available products in stock. */
  available?: InputMaybe<Scalars["Boolean"]["input"]>;
  /** The set of brand IDs which the result item brand ID must be included in. */
  brandId?: InputMaybe<Array<Scalars["Long"]["input"]>>;
  /** The set of category IDs which the result item category ID must be included in. */
  categoryId?: InputMaybe<Array<Scalars["Long"]["input"]>>;
  /** The set of EANs which the result item EAN must be included. */
  ean?: InputMaybe<Array<InputMaybe<Scalars["String"]["input"]>>>;
  /** An external parent ID or a list of IDs to search for products with the external parent ID. */
  externalParentId?: InputMaybe<Array<InputMaybe<Scalars["String"]["input"]>>>;
  /** Retrieve the product variant only if it contains images. */
  hasImages?: InputMaybe<Scalars["Boolean"]["input"]>;
  /** Ignores the display rules when searching for products. */
  ignoreDisplayRules?: InputMaybe<Scalars["Boolean"]["input"]>;
  /** Retrieve the product variant only if it is the main product variant. */
  mainVariant?: InputMaybe<Scalars["Boolean"]["input"]>;
  /** A parent ID or a list of IDs to search for products with the parent ID. */
  parentId?: InputMaybe<Array<Scalars["Long"]["input"]>>;
  /** The set of prices to filter. */
  prices?: InputMaybe<PricesInput>;
  /** The product unique identifier (you may provide a list of IDs if needed). */
  productId?: InputMaybe<Array<Scalars["Long"]["input"]>>;
  /** The product variant unique identifier (you may provide a list of IDs if needed). */
  productVariantId?: InputMaybe<Array<Scalars["Long"]["input"]>>;
  /** A product ID or a list of IDs to search for other products with the same parent ID. */
  sameParentAs?: InputMaybe<Array<Scalars["Long"]["input"]>>;
  /** Search products. */
  search?: InputMaybe<Array<InputMaybe<Scalars["String"]["input"]>>>;
  /** The set of SKUs which the result item SKU must be included. */
  sku?: InputMaybe<Array<InputMaybe<Scalars["String"]["input"]>>>;
  /** Show products with a quantity of available products in stock greater than or equal to the given number. */
  stock_gte?: InputMaybe<Scalars["Long"]["input"]>;
  /** Show products with a quantity of available products in stock less than or equal to the given number. */
  stock_lte?: InputMaybe<Scalars["Long"]["input"]>;
  /** The set of stocks to filter. */
  stocks?: InputMaybe<StocksInput>;
  /** The subscription group IDs to search for products inside those groups. */
  subscriptionGroupId?: InputMaybe<Array<Scalars["Long"]["input"]>>;
  /** Retrieve products which the last update date is greater than or equal to the given date. */
  updatedAt_gte?: InputMaybe<Scalars["String"]["input"]>;
  /** Retrieve products which the last update date is less than or equal to the given date. */
  updatedAt_lte?: InputMaybe<Scalars["String"]["input"]>;
};
// ...
export type ProductSortKeys =
  /** The applied discount to the product variant price. */
  | "DISCOUNT"
  /** The product name. */
  | "NAME"
  /** The product variant price. */
  | "PRICE"
  /** Sort in a random way. */
  | "RANDOM"
  /** The date the product was released. */
  | "RELEASE_DATE"
  /** The sales number on a period of time. */
  | "SALES"
  /** The quantity in stock of the product variant. */
  | "STOCK";
// ...
export type SortDirection =
  /** The results will be sorted in an ascending order. */
  | "ASC"
  /** The results will be sorted in an descending order. */
  | "DESC";
// ...
export type StocksInput = {
  /** The distribution center Ids to match. */
  dcId?: InputMaybe<Array<Scalars["Long"]["input"]>>;
  /** The distribution center names to match. */
  dcName?: InputMaybe<Array<InputMaybe<Scalars["String"]["input"]>>>;
  /** The product stock must be greater than or equal to. */
  stock_gte?: InputMaybe<Scalars["Long"]["input"]>;
  /** The product stock must be lesser than or equal to. */
  stock_lte?: InputMaybe<Scalars["Long"]["input"]>;
};
// ...
export type ProductFragment = {
  mainVariant?: boolean | null;
  productName?: string | null;
  productId?: any | null;
  alias?: string | null;
  available?: boolean | null;
  averageRating?: number | null;
  condition?: string | null;
  createdAt?: any | null;
  ean?: string | null;
  id?: string | null;
  minimumOrderQuantity?: number | null;
  productVariantId?: any | null;
  parentId?: any | null;
  sku?: string | null;
  numberOfVotes?: number | null;
  stock?: any | null;
  variantName?: string | null;
  variantStock?: any | null;
  collection?: string | null;
  urlVideo?: string | null;
  attributes?: Array<{
    id?: string | null;
    type?: string | null;
    value?: string | null;
    name?: string | null;
  } | null> | null;
  productCategories?: Array<{
    id: number;
    name?: string | null;
    url?: string | null;
    hierarchy?: string | null;
    main: boolean;
    googleCategories?: string | null;
  } | null> | null;
  informations?: Array<{
    title?: string | null;
    value?: string | null;
    type?: string | null;
  } | null> | null;
  images?: Array<{ url?: string | null; fileName?: string | null; print: boolean } | null> | null;
  prices?: {
    discountPercentage: any;
    discounted: boolean;
    listPrice?: any | null;
    multiplicationFactor: number;
    price: any;
    bestInstallment?: {
      discount: boolean;
      displayName?: string | null;
      fees: boolean;
      name?: string | null;
      number: number;
      value: any;
    } | null;
    installmentPlans?: Array<{
      displayName?: string | null;
      name?: string | null;
      installments?: Array<{
        discount: boolean;
        fees: boolean;
        number: number;
        value: any;
      } | null> | null;
    } | null> | null;
    priceTables?: Array<{
      discountPercentage: any;
      id: any;
      listPrice?: any | null;
      price: any;
    } | null> | null;
    wholesalePrices?: Array<{ price: any; quantity: number } | null> | null;
  } | null;
  productBrand?: {
    fullUrlLogo?: string | null;
    logoUrl?: string | null;
    name?: string | null;
    alias?: string | null;
  } | null;
  seller?: { name?: string | null } | null;
  similarProducts?: Array<{
    alias?: string | null;
    image?: string | null;
    imageUrl?: string | null;
    name?: string | null;
  } | null> | null;
  promotions?: Array<{
    content?: string | null;
    disclosureType?: string | null;
    id: any;
    fullStampUrl?: string | null;
    stamp?: string | null;
    title?: string | null;
  } | null> | null;
};
// ...
export type GetProductsQueryVariables = Exact<{
  filters: ProductExplicitFiltersInput;
  first: Scalars["Int"]["input"];
  sortDirection: SortDirection;
  sortKey?: InputMaybe<ProductSortKeys>;
  after?: InputMaybe<Scalars["String"]["input"]>;
  partnerAccessToken?: InputMaybe<Scalars["String"]["input"]>;
}>;

export type GetProductsQuery = {
  products?: {
    totalCount: number;
    nodes?: Array<{
      mainVariant?: boolean | null;
      productName?: string | null;
      productId?: any | null;
      alias?: string | null;
      available?: boolean | null;
      averageRating?: number | null;
      condition?: string | null;
      createdAt?: any | null;
      ean?: string | null;
      id?: string | null;
      minimumOrderQuantity?: number | null;
      productVariantId?: any | null;
      parentId?: any | null;
      sku?: string | null;
      numberOfVotes?: number | null;
      stock?: any | null;
      variantName?: string | null;
      variantStock?: any | null;
      collection?: string | null;
      urlVideo?: string | null;
      attributes?: Array<{
        id?: string | null;
        type?: string | null;
        value?: string | null;
        name?: string | null;
      } | null> | null;
      productCategories?: Array<{
        id: number;
        name?: string | null;
        url?: string | null;
        hierarchy?: string | null;
        main: boolean;
        googleCategories?: string | null;
      } | null> | null;
      informations?: Array<{
        title?: string | null;
        value?: string | null;
        type?: string | null;
      } | null> | null;
      images?: Array<{
        url?: string | null;
        fileName?: string | null;
        print: boolean;
      } | null> | null;
      prices?: {
        discountPercentage: any;
        discounted: boolean;
        listPrice?: any | null;
        multiplicationFactor: number;
        price: any;
        bestInstallment?: {
          discount: boolean;
          displayName?: string | null;
          fees: boolean;
          name?: string | null;
          number: number;
          value: any;
        } | null;
        installmentPlans?: Array<{
          displayName?: string | null;
          name?: string | null;
          installments?: Array<{
            discount: boolean;
            fees: boolean;
            number: number;
            value: any;
          } | null> | null;
        } | null> | null;
        priceTables?: Array<{
          discountPercentage: any;
          id: any;
          listPrice?: any | null;
          price: any;
        } | null> | null;
        wholesalePrices?: Array<{ price: any; quantity: number } | null> | null;
      } | null;
      productBrand?: {
        fullUrlLogo?: string | null;
        logoUrl?: string | null;
        name?: string | null;
        alias?: string | null;
      } | null;
      seller?: { name?: string | null } | null;
      similarProducts?: Array<{
        alias?: string | null;
        image?: string | null;
        imageUrl?: string | null;
        name?: string | null;
      } | null> | null;
      promotions?: Array<{
        content?: string | null;
        disclosureType?: string | null;
        id: any;
        fullStampUrl?: string | null;
        stamp?: string | null;
        title?: string | null;
      } | null> | null;
    } | null> | null;
    pageInfo: {
      hasNextPage: boolean;
      endCursor?: string | null;
      hasPreviousPage: boolean;
      startCursor?: string | null;
    };
  } | null;
};